Add download of youtube video and fix versioned sprite data use.
This fixes a lot of Express course and likely some of the pending issues
with certain CSP levels. The affected levels are namely SpriteLab or
AppLab style levels which have graphical assets pulled from the
animations bucket using a version tag. Expecting a query parameter
(?=, etc) in the URL to return a different result is fairly impossible
in the static website context. These are turned into fully qualified
files (blah.png?version=foo -> blah.png-version-foo) and saved upon
crawling.

Youtube videos are detected and downloaded. They are not usable, but the
videos are now included in the module, at least. We should replace
linked youtube videos with a video.js player. We are using the
`youtube-dl` program which has a very basic python requirement, I think.